pdlm-comparison-tool/requirements.txt
streamlit
pandas
openpyxl
python-docx
re